scrapper/requirements.txt
django
djangorestframework
celery
requests
selenium
redis
